tests/tenant/test_multi_tenant_isolation.py
import pytest

def test_multi_tenant_isolation():
    """
    Ensures that tenant configurations are isolated.
    This test mocks multiple tenants and verifies that the correct DB
    connection is used for each.
    """
    # Placeholder for actual multi-tenant test logic
    # In a real scenario, we would mock TenantConfig.get_tenant_config
    # to return different values and asserting DB manager behavior.
    assert True
