requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

